improvements to location input on event types
- extracted all location logic from event types page
- prevented modal for showing unless LocationType.InPerson as that requires an extra address input
- removed SVGs from event types code
- TODO: big bad bug after adding more than one location
